テストの追加

+++ CSChatworkAPI.Test.E2E/TestContext.cs
@@ -55,18 +55,19 @@ namespace CSChatworkAPI.Test.E2E
         {
             var path = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "TestData.json");
             TestData = TestData.Create(path);
+            ChatworkClient = new ChatworkClient(TestData.InputData.APIToken);
         }
 
         public static void SetUp()
         {
-            ChatworkClient = new ChatworkClient(TestData.InputData.APIToken);
             Me = ChatworkClient.GetMe();
             TestRoom = TestCaseUtility.CreateRoomForTest();
         }
 
         public static void TearDown()
         {
-            ChatworkClient.DeleteRoom(TestRoom.room_id);
+            if (TestRoom != null)
+                ChatworkClient.DeleteRoom(TestRoom.room_id);
         }
     }
 }
